fix(kotlin): move jni_emitter tests last in the flattened module
jni_emitter.rs include!s its sub-files into one module. bridge_object.rs
(included first) ended with the `#[cfg(test)] mod tests`, so every production
item in the seven files included after it counted as appearing after a test
module — tripping clippy::items_after_test_module under -D warnings. Extract
the test module into jni_emitter/tests.rs and include! it last so the test
module is the final item. No behavior change; the test still passes.
