+++ src/backends/kotlin/gen_bindings/jni_emitter/bridge_object.rs
@@ -237,29 +237,3 @@ fn trait_bridge_manages_jni_function(func_name: &str, config: &ResolvedCrateConf
                 || bridge.clear_fn.as_deref() == Some(func_name))
     })
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::core::config::{KotlinAndroidConfig, TraitBridgeConfig};
-
-    #[test]
-    fn jni_bridge_object_treats_android_trait_lifecycle_functions_as_managed() {
-        let config = ResolvedCrateConfig {
-            kotlin_android: Some(KotlinAndroidConfig::default()),
-            trait_bridges: vec![TraitBridgeConfig {
-                trait_name: "Renderer".to_string(),
-                register_fn: Some("register_renderer".to_string()),
-                unregister_fn: Some("unregister_renderer".to_string()),
-                clear_fn: Some("clear_renderers".to_string()),
-                ..TraitBridgeConfig::default()
-            }],
-            ..ResolvedCrateConfig::default()
-        };
-
-        assert!(trait_bridge_manages_jni_function("register_renderer", &config));
-        assert!(trait_bridge_manages_jni_function("unregister_renderer", &config));
-        assert!(trait_bridge_manages_jni_function("clear_renderers", &config));
-        assert!(!trait_bridge_manages_jni_function("list_renderers", &config));
-    }
-}
